+++ lib/classes/class-bootstrap.php
@@ -282,7 +282,7 @@ namespace UsabilityDynamics\WPP {
           try {
 
             // Must be able to parse composer.json from plugin file, hopefully to detect the "_build.sha" field.
-            $_composer = json_decode( @file_get_contents( ABSPATH . '/'  . $_product_path . '/composer.json' )  );
+            $_composer = json_decode( @file_get_contents(  trailingslashit( WPP_Path ) .'/composer.json' )  );
 
             if( is_object( $_composer ) && $_composer->extra && isset( $_composer->extra->_build ) && isset( $_composer->extra->_build->sha ) ) {
               $_version = $_composer->extra->_build->sha;