+++ changes.md
@@ -1,4 +1,5 @@
 ### 2.2.1 
+* Improved UX for viewing property objects by using a JSON viewer library.
 * Improved wp-rets-client logic for inserting terms and taxonomies. Both are added on-the-fly, with hierarchy support if needed.
 * Added Feature flags to help tab to help understand what features are enabled.
 * Fixed a bug with rets-client site registration hanging.